+++ packages/backend/src/core/CoreModule.ts
@@ -38,6 +38,7 @@ import { IdService } from './IdService.js';
 import { ImageProcessingService } from './ImageProcessingService.js';
 import { MiniAppManifestService } from './MiniAppManifestService.js';
 import { MiniAppOAuthTokenService } from './MiniAppOAuthTokenService.js';
+import { UserMiniAppService } from './UserMiniAppService.js';
 import { SystemAccountService } from './SystemAccountService.js';
 import { InternalStorageService } from './InternalStorageService.js';
 import { MetaService } from './MetaService.js';
@@ -339,6 +340,7 @@ const $ApQuestionService: Provider = { provide: 'ApQuestionService', useExisting
 		ImageProcessingService,
 		MiniAppManifestService,
 		MiniAppOAuthTokenService,
+		UserMiniAppService,
 		InternalStorageService,
 		MetaService,
 		MfmService,
@@ -643,6 +645,7 @@ const $ApQuestionService: Provider = { provide: 'ApQuestionService', useExisting
 		ImageProcessingService,
 		MiniAppManifestService,
 		MiniAppOAuthTokenService,
+		UserMiniAppService,
 		InternalStorageService,
 		MetaService,
 		MfmService,